chore(logger): make default level to INFO
* Update riklet, rikctl and controller loggers to have INFO as default
  level
* Remove traits and utils rust files from riklet to move them into
  main.rs, to remove the number of modules
